+++ compiler/hash-fmt/src/state.rs
@@ -5,12 +5,12 @@
 #[derive(Debug, Clone, Default)]
 pub struct AstPrinterState {
     /// The current level of the indentation.
-    pub indentation: u8,
+    pub indentation: u32,
 
     /// The current line width of the line, this is used to control whether the
     /// formatter should wrap the current line where it can or if it can
     /// continue to print some item on the current line.
-    pub width: u16,
+    pub width: u32,
 
     /// The buffer that is used to store the result of the formatting
     /// temporarily until the entire file or AST node is ready to be printed.
